requirements.txt
python-fasthtml
matplotlib
numpy
scikit-learn

